Static pages/styling
Allowed for static pages for CSS/JS/IMAGES within the django project in the settings.py. Created static folder to hold the folders for css/js/images. Adjusted the styling for the nav bar to be the same red as the footer and a darker shade when hovering and when on a page.